+++ Lib/core/index.js
@@ -2,11 +2,6 @@ const { ReactStateManagement, VueStateManagement, AngularStateManagement } = req
 const Parser = require("./JavascriptManagement");
 
 /**
- * Array of Main Component's Components.
- */
-var Components = [];
-
-/**
  * React Compiler
  * 
  * Function that take the HTML string and translate to React
@@ -38,8 +33,6 @@ const ReactCompiler = (name, html, css) => {
 
 	RStateManagement.setVarsToStatesContent(parse.vars);
 
-	Components = RStateManagement.componentsContent;
-
 	//Add new lines and idents to code beauty
 	let pretty = RStateManagement
 		.filterHTML(html)
@@ -63,11 +56,14 @@ ${pretty}\t\t)
 }
 export default ${name || "MyComponent"};
 `;
-	return template;
+	return {
+		main:template,
+		components:RStateManagement.componentsContent
+	};
 };
 
 /**
- * Vue Translator
+ * Vue Compiler
  * 
  * Function that take the HTML string and translate to Vue
  *
@@ -81,6 +77,7 @@ const VueCompiler = (name, html, css) => {
 
 	let VStateManagement = new VueStateManagement();
 
+	
 	let parse = new Parser(); //JS Parser
 
 	// Set Styles
@@ -102,8 +99,6 @@ const VueCompiler = (name, html, css) => {
 	
 	VStateManagement.setVarsToStatesContent(parse.vars);
 	
-	Components = VStateManagement.componentsContent;
-	
 	//Add new lines and idents to code beauty
 	let pretty = VStateManagement
 		.filterHTML(html)
@@ -119,11 +114,14 @@ const VueCompiler = (name, html, css) => {
 ${VStateManagement.componentData(name)}
 ${style}`;
 
-	return component;
+	return {
+		main:component,
+		components:VStateManagement.componentsContent
+	};
 };
 
 /**
- * Angular Translator
+ * Angular Compiler
  * 
  * Function that take the HTML string and translate to Angular
  *
@@ -133,52 +131,57 @@ ${style}`;
  * 
  * @return {string}
  */
-const AngularCompiler = (name, html, css, js) => {
+const AngularCompiler = (name, html, css) => {
 
-	let AStateManagement = new AngularStateManagement();
+	const AStateManagement = new AngularStateManagement();
 
-	let parse = new Parser(js); //JS Parser
+	const parse = new Parser(); //JS Parser
 
+	// Set Styles
 	let style = css !== "" ? `<style scoped>\n${css}</style>` : "";
-
-	//Get all data from HTML string
-	AStateManagement.getHTMLString(html);
 
 	//Get states declarations from JS and set to Data
 	AStateManagement.statesFromJS = parse.states;
 
+	//Parse Lifecycles
+	AStateManagement.setLifecycle(parse.lifecycles, "a");
+
+	//Get all data from HTML string
+	AStateManagement.getHTMLString(html);
+
 	//Get Methods from JS and set to Data
-	AStateManagement.getJsData(parse.functions, "v");
-
+	AStateManagement.getJsData(parse.functions, "a");
+	
 	AStateManagement.watchers = parse.watchers;
-
+	
 	AStateManagement.setVarsToStatesContent(parse.vars);
 	
-	Components = AStateManagement.componentsContent;
-	/*
-Component Template
+	//Add new lines and idents to code beauty
+	let pretty = AStateManagement
+		.filterHTML(html)
+		.split(/\n/)
+		.map(e => {
+			if (e) return `\t${e}\n`;
+		})
+		.join("");
 
+	const component = `import { Component${AStateManagement.props.length > 0 ? ", Input" : ""}} from '@angular/core';
+${AStateManagement.components.map(e => `\nimport { ${e} } from "./components/${e}";`).join("")}
 
-import { Component } from '@angular/core';
- 
 @Component({
-  selector: 'app-root',
-  template: `
-    <h1>{{title}}</h1>
-    <h2>My favorite hero is: {{myHero}}</h2>
-    `
+	selector: '${AStateManagement.generateComponentName(name)}-root',
+	template:\`${pretty}\`
 })
-export class AppComponent {
-  title = 'Tour of Heroes';
-  myHero = 'Windstorm';
-}
 
-*/
-	console.log(AStateManagement.componentData);
-	process.exit(1);
+export class ${name} {
+	${AStateManagement.componentData}
+}`;
+	return {
+		main:component,
+		components:AStateManagement.componentsContent
+	};
 };
 
-exports.Components = Components;
 exports.VueCompiler = VueCompiler;
 exports.AngularCompiler = AngularCompiler;
 exports.ReactCompiler = ReactCompiler;