+++ bin/changelog.js
@@ -1,4 +1,4 @@
 const clc = require("cli-color");
 
 console.log(clc.greenBright(`Congratulations!!!\n\n`) + clc.whiteBright("We have new things!!!\n"));
-console.log("If you want know more, go to: " + clc.whiteBright(`https://github.com/Davids-Devel/rocket-translator/blob/master/README.md`));
+console.log("If you want know more, go to: " + clc.whiteBright(`https://github.com/Davids-Devel/rocket-translator/blob/master/CHANGELOG.md`));